+++ mhu-lib-core/src/main/java/de/mhus/lib/core/MDate.java
@@ -341,7 +341,7 @@ public class MDate extends Date {
 
 	public static Date toDateOnly(Date date) {
 		if (date == null) return null;
-		return new Date( date.getTime() / MTimeInterval.DAY_IN_MILLISECOUNDS * MTimeInterval.DAY_IN_MILLISECOUNDS );
+		return new Date( date.getTime() / MPeriod.DAY_IN_MILLISECOUNDS * MPeriod.DAY_IN_MILLISECOUNDS );
 	}
 
 	/**