+++ mhu-lib-core/src/main/java/de/mhus/lib/core/MStopWatch.java
@@ -126,7 +126,7 @@ public class MStopWatch extends MJmx {
 	}
 
 	public String getCurrentTimeAsString() {
-		return MTimeInterval.getIntervalAsString(getCurrentTime());
+		return MPeriod.getIntervalAsString(getCurrentTime());
 	}
 
 	public String getName() {