+++ mhu-lib-persistence/src/main/java/de/mhus/lib/adb/transaction/MemoryLockStrategy.java
@@ -19,12 +19,12 @@ import java.util.HashMap;
 
 import de.mhus.lib.adb.Persistable;
 import de.mhus.lib.core.MThread;
-import de.mhus.lib.core.MTimeInterval;
+import de.mhus.lib.core.MPeriod;
 import de.mhus.lib.errors.TimeoutRuntimeException;
 
 public class MemoryLockStrategy extends LockStrategy {
 
-	private long maxLockAge = MTimeInterval.HOUR_IN_MILLISECOUNDS;
+	private long maxLockAge = MPeriod.HOUR_IN_MILLISECOUNDS;
 	private long sleepTime = 200;
 
 	private HashMap<String, LockObject> locks = new HashMap<>();