+++ docs/release-notes.md
@@ -5,3 +5,4 @@
 * Removed Lamda Function Hack (not Java 12 compatible)
 * Add const gnerator plugin for maven to use field references
 * DB Query generator use integer counter for variable place holders instead of UUID (improve readability of log files)
+* Reorganize MTimeInterval and rename to MPeriod